MAINT: remove uncommitted tutorials
#62 and #75 rightly point out that there are missing tutorials.
These tutorials were pushed out for the [neuro-symbolic summer school '22](https://ibm.github.io/neuro-symbolic-ai/events/ns-summerschool2022/) as part of the badge content. It was decided that the rest of the notebooks should be released as part of a second badge - since understanding first order logic (and specifically the LNN implementation thereof) is more complex than a fundamental level. 
I'll open up a new branch so that others can work on the tutorial content for the second badge and place the (currently incomplete) notebooks 3-10 in there